+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
 
-import Home from "./components/Home";
+import Home from "./pages/Home";
 import AboutUs from "./components/AboutUs";
 import Layout from "./components/Layout";
 import Login from "./pages/Login";
